feat: add test framework, address method and server listen/close/error callbacks
* Adds mocha test framework
* Adds `.address()` method to `Http3Server` to know which port/host it ends up listening on
* Adds `.ready` and `closed` promises to `Http3Server` so you don't have to guess when it's ready to use

The `.address()` method is necessary to know which port a server listens on when you specify `0` as the parameter to avoid clashes during testing.

The `.ready` and `.closed` promises are triggered by callbacks from the C++ code - I hope I implemented these correctly.

I've replicated all of the tests from the current suite, these can be deleted in a follow up PR.

Because we don't have to use an arbitrary timeout to wait for the server to be ready the test suite is now much faster to run - on my machine the port is opened
in 20-30ms so no need to wait 2000ms as in the current suite.

The event loop still takes a long time to shut down though, maybe this can be refactored?

Fixes #100