fix: build types before publish
The `dist` folder needs to be built in order to include types in
the tarball published on npm so I've added a `Build types` step
to the two publish jobs after install but before publish.

This might be better in the `prebuild` subcommand in `build.js`?
Not sure, we don't want to end up building types over and over again.